Added new moon icons and set them to change based on the moon phase. They are based in the grid and the overview and change accordingly.
Created new js function files for displaying the moonphase icon and name of moonphase.

Also added the named day of the week above the date.

Changed the styling of the location inputs.

Added a new Font from Google called Outline, however this may be temporary as I'm unsure of the design.