Updated the styling.
Changed font to Nunito Sans.

Adjusted the layout of the moon info panel and added margin beneath this for separate the grid.

Adjusted the styling for the location input.

Added precipitation (mm) param, as the probability isn't available on the UKMO weather model. Added functionality to change between mm and % based on weather model, and updated the grid colour to match.