Added several responsive changes for desktop.
Changed the layout of the mobile menu again to a card view that has a fixed position at the bottom of the screen. Added a close button to the menu for easy access when holding a phone.

Added a red light filter.

Created a desktop nav component and rendered the toggle buttons. Added functionality that switched between mobile and desktop nav.

Changed the layout of the location input on desktop.